+++ pkg/config/snapshot_config.go
@@ -1,5 +1,9 @@
 package config
 
+import (
+	flag "github.com/spf13/pflag"
+)
+
 const (
 	// which snapshot type to load. 'local' or 'global'
 	CfgSnapshotLoadType = "snapshots.loadType"
@@ -28,20 +32,20 @@ const (
 )
 
 func init() {
-	NodeConfig.SetDefault(CfgSnapshotLoadType, "local")
-	NodeConfig.SetDefault(CfgLocalSnapshotsDepth, 50)
-	NodeConfig.SetDefault(CfgLocalSnapshotsIntervalSynced, 50)
-	NodeConfig.SetDefault(CfgLocalSnapshotsIntervalUnsynced, 1000)
-	NodeConfig.SetDefault(CfgLocalSnapshotsPath, "export.bin")
-	NodeConfig.SetDefault(CfgLocalSnapshotsDownloadURL, "")
-	NodeConfig.SetDefault(CfgGlobalSnapshotPath, "snapshotMainnet.txt")
-	NodeConfig.SetDefault(CfgGlobalSnapshotSpentAddressesPaths, []string{
+	flag.String(CfgSnapshotLoadType, "local", "which snapshot type to load. 'local' or 'global'")
+	flag.Int(CfgLocalSnapshotsDepth, 50, "the depth, respectively the starting point, at which a local snapshot of the ledger is generated")
+	flag.Int(CfgLocalSnapshotsIntervalSynced, 50, "interval, in milestone transactions, at which snapshot files are created if the ledger is fully synchronized")
+	flag.Int(CfgLocalSnapshotsIntervalUnsynced, 1000, "interval, in milestone transactions, at which snapshot files are created if the ledger is not fully synchronized")
+	flag.String(CfgLocalSnapshotsPath, "snapshots/mainnet/export.bin", "path to the local snapshot file")
+	flag.String(CfgLocalSnapshotsDownloadURL, "", "URL to load the local snapshot file from")
+	flag.String(CfgGlobalSnapshotPath, "snapshotMainnet.txt", "path to the global snapshot file containing the ledger state")
+	flag.StringSlice(CfgGlobalSnapshotSpentAddressesPaths, []string{
 		"previousEpochsSpentAddresses1.txt",
 		"previousEpochsSpentAddresses2.txt",
 		"previousEpochsSpentAddresses3.txt",
-	})
-	NodeConfig.SetDefault(CfgGlobalSnapshotIndex, 1050000)
-	NodeConfig.SetDefault(CfgPruningEnabled, true)
-	NodeConfig.SetDefault(CfgPruningDelay, 40000)
-	NodeConfig.SetDefault(CfgSpentAddressesEnabled, true)
+	}, "paths to the spent addresses files")
+	flag.Int(CfgGlobalSnapshotIndex, 1050000, "milestone index of the global snapshot")
+	flag.Bool(CfgPruningEnabled, true, "whether to delete old transaction data from the database")
+	flag.Int(CfgPruningDelay, 40000, "amount of milestone transactions to keep in the database")
+	flag.Bool(CfgSpentAddressesEnabled, true, "enable support for wereAddressesSpentFrom (needed for Trinity, but local snapshots are much bigger)")
 }