+++ pkg/model/tangle/tags_storage.go
@@ -80,13 +80,7 @@ func StoreTag(txTag hornet.Hash, txHash hornet.Hash) *CachedTag {
 
 	tag := hornet.NewTag(txTag[:17], txHash[:49])
 
-	cachedObj := tagsStorage.ComputeIfAbsent(tag.ObjectStorageKey(), func(key []byte) objectstorage.StorableObject { // tag +1
-		tag.Persist()
-		tag.SetModified()
-		return tag
-	})
-
-	return &CachedTag{CachedObject: cachedObj}
+	return &CachedTag{CachedObject: tagsStorage.Store(tag)}
 }
 
 // tag +-0