+++ plugins/tangle/solidifier.go
@@ -287,8 +287,6 @@ func solidifyMilestone(newMilestoneIndex milestone.Index, force bool) {
 	- Everytime a request queue gets empty, start the solidifier for the next known non-solid milestone
 	- If tx are missing, they are requested by the solidifier
 	- The traversion can be aborted with a signal and restarted
-	- If we miss more than WARP_SYNC_THRESHOLD milestones in our requests, request them via warp sync
-
 	*/
 
 	if !force {
@@ -297,13 +295,23 @@ func solidifyMilestone(newMilestoneIndex milestone.Index, force bool) {
 				- newMilestoneIndex==0 (triggersignal) and solidifierMilestoneIndex==0 (no ongoing solidification)
 				- newMilestoneIndex==solidMilestoneIndex+1 (next milestone)
 				- newMilestoneIndex!=0 (new milestone received) and solidifierMilestoneIndex!=0 (ongoing solidification) and newMilestoneIndex<solidifierMilestoneIndex (milestone older than ongoing solidification)
+				- newMilestoneIndex!=0 (new milestone received) and solidifierMilestoneIndex==0 (no ongoing solidification) and RequestQueue().Empty() (request queue is already empty)
+
+			The following events trigger the solidifier in the node:
+				- new valid milestone was processed (newMilestoneIndex=index, force=false)
+				- a milestone was missing in the cone at solidifier run (newMilestoneIndex=0, force=true)
+				- WebAPI call (newMilestoneIndex=0, force=true)
+				- milestones in warp sync range were already in database at warpsync startup (newMilestoneIndex==0, force=true)
+				- another milestone was successfully solidified (newMilestoneIndex=0, force=false)
+				- request queue gets empty and node is not synced (newMilestoneIndex=0, force=true)
 		*/
 
 		solidifierMilestoneIndexLock.RLock()
 		triggerSignal := (newMilestoneIndex == 0) && (solidifierMilestoneIndex == 0)
 		nextMilestoneSignal := newMilestoneIndex == tangle.GetSolidMilestoneIndex()+1
 		olderMilestoneDetected := (newMilestoneIndex != 0) && ((solidifierMilestoneIndex != 0) && (newMilestoneIndex < solidifierMilestoneIndex))
-		if !(triggerSignal || nextMilestoneSignal || olderMilestoneDetected) {
+		newMilestoneReqQueueEmptySignal := (solidifierMilestoneIndex == 0) && (newMilestoneIndex != 0) && gossip.RequestQueue().Empty()
+		if !(triggerSignal || nextMilestoneSignal || olderMilestoneDetected || newMilestoneReqQueueEmptySignal) {
 			// Do not run solidifier
 			solidifierMilestoneIndexLock.RUnlock()
 			return
@@ -408,11 +416,11 @@ func getConfirmedMilestoneMetric(cachedMsTailTx *tangle.CachedTransaction, miles
 	newMilestoneTimestamp := time.Unix(cachedMsTailTx.GetTransaction().GetTimestamp(), 0)
 	cachedMsTailTx.Release()
 
-	oldMilestone := tangle.GetCachedMilestoneOrNil(milestoneIndexToSolidify - 1)
+	oldMilestone := tangle.GetCachedMilestoneOrNil(milestoneIndexToSolidify - 1) // milestone +1
 	if oldMilestone == nil {
 		return nil, ErrMilestoneNotFound
 	}
-	defer oldMilestone.Release(true)
+	defer oldMilestone.Release(true) // milestone -1
 
 	oldMilestoneTailTx := tangle.GetCachedTransactionOrNil(oldMilestone.GetMilestone().Hash)
 	if oldMilestoneTailTx == nil {